img/py/requirements_windows.txt
accessible-pygments==0.0.5
adaptive==1.4.1
addcopyfighandler==3.2.4
alabaster==1.0.0
asttokens==3.0.0
attrs==25.3.0
autobahn==24.4.2
babel==2.17.0
beautifulsoup4==4.13.4
broadbean==0.14.0
certifi==2025.8.3
cf-xarray==0.10.7
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
colorspacious==1.1.2
comm==0.2.3
contourpy==1.3.3
crc32c==2.7.1
cryptography==45.0.6
cycler==0.12.1
dask==2025.7.0
dataclassabc==0.0.18
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
docutils==0.21.2
donfig==0.8.1.post1
et-xmlfile==2.0.0
executing==2.2.0
ezdxf==1.4.2
fastz==1.0.0
filter-functions==1.1.3
fonttools==4.59.1
fsspec==2025.7.0
gitdb==4.0.12
gitpython==3.1.45
h5netcdf==1.6.4
h5py==3.14.0
hdf5storage==0.2.0
hickle==5.0.3
hyperlink==21.0.0
idna==3.10
imagecodecs==2025.8.2
imageio==2.37.0
imageio-ffmpeg==0.6.0
imagesize==1.4.1
importlib-metadata==8.7.0
ipykernel==6.30.1
ipython==8.37.0
ipywidgets==8.1.7
jax==0.7.0
jaxlib==0.7.0
jedi==0.19.2
jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-client==8.6.3
jupyter-core==5.8.1
jupyterlab-widgets==3.0.15
kerchunk==0.2.9
kiwisolver==1.4.9
lazy-loader==0.4
-e file:///Z:/Code/lindblad_mc_tools
line-profiler==5.0.0
llvmlite==0.44.0
locket==1.0.0
loky==3.5.5
lxml==6.0.0
markdown-it-py==3.0.0
markupsafe==3.0.2
matlabengine==25.1.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdit-py-plugins==0.5.0
mdurl==0.1.2
mjolnir @ git+https://git-ce.rwth-aachen.de/qutech/lab_software/python-mjolnir.git@b8dc7fe777e8b90c24575283b0a94fefab8c3726
ml-dtypes==0.5.3
more-itertools==10.7.0
moviepy==2.2.1
mpmath==1.3.0
msl-loadlib==1.0.0
myst-parser==4.0.1
nest-asyncio==1.6.0
networkx==3.5
noisyopt==0.2.2
numcodecs==0.16.2
numpy-quaternion==2024.0.10
openpyxl==3.1.5
opentelemetry-api==1.36.0
opt-einsum==3.4.0
packaging==25.0
pandas==2.3.1
parso==0.8.4
partd==1.4.2
pillow==11.3.0
platformdirs==4.3.8
poisson-schroedinger @ git+https://git-ce.rwth-aachen.de/qutech/simulation/poisson_schroedinger_1d.git@34dad33bac59e9ae2a2096c6720ee784dbfa6c0e
proglog==0.1.12
prompt-toolkit==3.0.51
psutil==7.0.0
pure-eval==0.2.3
pyarrow==21.0.0
pycparser==2.22
pydata-sphinx-theme==0.16.1
pygments==2.19.2
pymoosh==3.92
pyparsing==3.2.3
pyserial==3.5
pyside6==6.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-spectrometer==2025.7.1
pytz==2025.2
pyvisa==1.15.0
pyvisa-py==0.8.0
pywin32==311
pyyaml==6.0.2
pyzmq==27.0.1
qcodes @ git+https://github.com/qutech/qcodes.git@7f5c71f316c1a2c4b009f6ff2f46ba15c286afbc
qcodes-contrib-drivers @ git+https://github.com/qutech-lab/qcodes_contrib_drivers.git@192871bb5cb353cc563fbd4513c2d3358ce5bbda
qcodes-loop==0.1.3
qopt==1.3.5
qutech-util==2025.8.1
qutip==5.2.0
referencing==0.36.2
refractiveindex==0.0.7
requests==2.32.4
roman-numerals-py==3.1.0
rpds-py==0.27.0
ruamel-yaml==0.18.14
ruamel-yaml-clib==0.2.12
schema==0.7.7
scipy==1.16.1
setuptools==80.9.0
shiboken6==6.9.1
six==1.17.0
smmap==5.0.2
snowballstemmer==3.0.1
sortedcollections==2.1.0
sortedcontainers==2.4.0
soupsieve==2.7
sparse==0.17.0
sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
spyder-kernels==3.0.5
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tifffile==2025.6.11
toolz==1.0.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
txaio==25.6.1
typeguard==4.4.4
typing-extensions==4.14.1
tzdata==2025.2
ujson==5.10.0
uncertainties==3.2.3
urllib3==2.5.0
versioningit==3.3.0
wcwidth==0.2.13
websockets==15.0.1
wheel==0.45.1
widgetsnbextension==4.0.14
wrapt==1.17.3
xarray==2025.8.0
zarr==3.1.1
zipp==3.23.0
